backend/requirements.txt
flask
libvirt-python
requests
websockify
psutil
